Fix consistency of missing thumbnail size
Unavailable thumbnails use a 120x90 placeholder, which doesn't quite
span the same width as a "full" thumbnail.  This normally shouldn't
matter much in mixed {un,}available thumbnails since the existence
of the latter will cause the former to fit the same column width.

But with searching, it's now possible to get a set of results that
only show the 'missing' thumbnail, causing the column to resize.

This change now sets a fixed width instead of maximum height.  This
should result in approximately the same final appearance.