Set default options for ProxyFix to enable secure websockets as needed
This doesn't seem ideal, but it is probably the default we need to
use, since I don't think url_for normally needs to care about what
protocol is in use (since it usually only renders paths anchored
to the web root, so it cares not about HTTP/S).

We may opt to pass _scheme into url_for as part of the config
instead, which would probably be more reasonable.
I suppose we'll worry about that later.