Only add finished matches to ignore history
Feed monitoring now only adds video IDs to the cache for matches that
were seen and known to be finalized VODs.

This means that matches will always be scheduled if they are upcoming,
live, or recently finished (if they aren't already queued).

And a player request is only made for those matches if, again, they
aren't already in the queue.

This partially fixes previously scheduled streams being lost on
restart, since at this time only finished or errored streams are
stored.  Though in the future it'd be good if active tasks could be
recovered on restart as well.

Closes #4.